Fix unused braces warnings in macro expansion
Previously, `From` implementations may have expanded to code similar to the
following:

```
...
where
    [(); { Size }]:
        ::modular_bitfield::private::IsU32Compatible,
...
```

Where `Size` is a constant expression that reduces to a `usize`.
With `Size` being an expression and the context in which it is used, the braces
around it are unneeded, producing a warning.

The braces are inserted in `impl::bitfield::expand::generate_bitfield_size`,
surrounding the returned sum-expression.

To avoid the warning, `generate_bitfield_size` was modified to return a naked
expression.

Places where brackets were needed to avoid an incorrect reading of the
expression were modified to add the brackets in place.
Similarly, places where the brackets were not needed were simplified.

Fixes Robbepop/modular-bitfield#66.
Closes Robbepop/modular-bitfield#68.