Reverts the parenthesization of generate_bitfield_size.
Recently, `generate_bitfield_size` was modified to return an expression not
contained in a block, to avoid generating an `unused_braces` warning.
See: https://github.com/Robbepop/modular-bitfield/issues/66

To simplify the reading of the generated code and avoid ambiguities with
operators' precedence, further parenthesization was added to the expression
itself.

The new parenthesization generates an `unused_parenthesis` warning and is now removed.