Avoids an incorrect evaluation of `expand_byte_conversion_impls`.
After the recent changes to `generate_bitfield_size`, the returned expression is
not enclosed in a block.

In `expand_byte_conversion_impls`, when `filled` is not enabled, an expression
of the following form is generated:

```
if bytes[...#next_divisible_by_8 - #size)))
```

Where `#size` is a value returned by `generate_bitfield_size`.
Since `#size` may take the form of a sum, it is now possible to generate an
expression of the form `#next_divisible_by_8 - size_1 + size_2 + ... + size_n`
which is not equivalent to the previous behavior, where `#size` was first
reduced to a single value and the subtracted, and results in a bigger result.

To avoid this issue, `#size` is now parenthesized in the expression to perform
the correct reduction.