Fix shift overflow checking byte bounds
A proper fix here would be to omit the check entirely when there
is no way that the value could be out of bounds because the struct
bitsize is a multiple of 8, but at this point in the generation
things are already converted into token streams and I am not
familiar enough with the codebase to understand where to introduce
such a test. I am only trying to un-bitrot this code enough to
make tests pass again, so this workaround should be OK for now.