Un-bitrot the crate (#108)
* Update to syn 2

* Silence warning about unused braces

This is caused by braces around `#actual_bits`. Removing
these requires extra work since they are required in other
places, so just suppress the warning.

* Fix shift overflow checking byte bounds

A proper fix here would be to omit the check entirely when there
is no way that the value could be out of bounds because the struct
bitsize is a multiple of 8, but at this point in the generation
things are already converted into token streams and I am not
familiar enough with the codebase to understand where to introduce
such a test. I am only trying to un-bitrot this code enough to
make tests pass again, so this workaround should be OK for now.

* Update trybuild tests

* Update GitHub workflow

* Replaces abandoned actions with maintained ones
* Reduce time spent recompiling and spinning up new instances
* Lint all targets
* Add documentation builds

* Fix new clippy lints

* Update to Rust edition 2021

* Use standard rustfmt features

Upstream abandoned rustfmt 2.0 several years ago and do not seem to
have any momentum toward stabilising many new features, so trying
to continue to use unstable features does not make sense when they
are unlikely to ever be broadly usable by other contributors.

* Update dev-dependencies to latest versions